rename package to be non-eclipse specific
rename package to be non-eclipse specific
