+++ pom.xml
@@ -2,15 +2,14 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>com.google.eclipse.gpe</groupId>
-    <artifactId>trunk</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <relativePath>../../</relativePath>
-  </parent>
-  <groupId>com.google.eclipse.gpe</groupId>
-  <artifactId>com.google.gdt.eclipse.login</artifactId>
+
+  <groupId>com.google.cloud.tools.ide.login</groupId>
+  <artifactId>com.google.cloud.tools.ide.login</artifactId>
   <version>0.0.0-SNAPSHOT</version>
+
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+  </properties>
 
   <dependencies>
     <dependency>
@@ -31,4 +30,3 @@
   </dependencies>
 
 </project>
-
