+++ src/main/java/com/google/cloud/tools/ide/login/OAuthDataStore.java
@@ -12,26 +12,32 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  *******************************************************************************/
-package com.google.gdt.eclipse.login.common;
+package com.google.cloud.tools.ide.login;
 
 /**
- * Presents a common API, implementable on a variety of platforms, for making log entries.
+ * Presents a common API, implementable on a variety of platforms, for storing a particular user's
+ * {@link OAuthData} object persistently, retrieving it, and clearing it.
  */
-public interface LoggerFacade {
-
+public interface OAuthDataStore {
+  
   /**
-   * Create an error log entry with a specified message and a specified {@link Throwable}.
+   * Stores a specified {@link OAuthData} object persistently.
    * 
-   * @param message the specified message
-   * @param t the specified {@code Throwable}
+   * @param credentials the specified {@code Credentials object}
    */
-  void logError(String message, Throwable t);
-
+  void saveOAuthData(OAuthData credentials);
+  
   /**
-   * Create a warning log entry with a specified message.
+   * Retrieves the persistently stored {@link OAuthData} object, if any.
    * 
-   * @param message the specified message
+   * @return
+   *     if there is a persistently stored {@code OAuthData} object, that object; otherwise an
+   *     {@code OAuthData} object all of whose getters return {@code null}
    */
-  void logWarning(String message);
-
+  OAuthData loadOAuthData();
+  
+  /**
+   * Clears the persistently stored {@link OAuthData} object, if any.
+   */
+  void clearStoredOAuthData();
 }
